Add support for complex numbers and fix type handling

+++ pkg/injector/map.go
@@ -100,20 +100,15 @@ func makeMapKey(keyType reflect.Type, entry *control.Entry) (reflect.Value, erro
 	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
 		mapKey = reflect.ValueOf(int(entry.GetCurrentIndex().GetInt64()))
 	case reflect.String:
-		// get the index as a string
 		mapKey = reflect.ValueOf(entry.GetCurrentIndex().GetString_())
-		// TODO: Handle below casese
 	case reflect.Bool:
 		mapKey = reflect.ValueOf(entry.GetCurrentIndex().GetBool())
 	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		mapKey = reflect.ValueOf(int(entry.GetCurrentIndex().GetUint64()))
+		mapKey = reflect.ValueOf(uint(entry.GetCurrentIndex().GetUint64()))
 	case reflect.Float32:
-		mapKey = reflect.ValueOf(int(entry.GetCurrentIndex().GetFloat32()))
+		mapKey = reflect.ValueOf(float32(entry.GetCurrentIndex().GetFloat32()))
 	case reflect.Float64:
-		mapKey = reflect.ValueOf(int(entry.GetCurrentIndex().GetFloat64()))
-	// case reflect.Pointer:
-	// case reflect.Interface:
-	// case reflect.Struct:
+		mapKey = reflect.ValueOf(entry.GetCurrentIndex().GetFloat64())
 	default:
 		return reflect.Value{}, fmt.Errorf("cannot create key of type %s", keyType.Kind())
 	}
